+++ packages/transitions/src/index.ts
@@ -1,1 +1,2 @@
 export * from './lib/Collapse';
+export * from './lib/Fade';