+++ packages/page-header/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.2.36](https://github.com/Availity/element/compare/@availity/mui-page-header@0.2.35...@availity/mui-page-header@0.2.36) (2024-06-27)
+
+### Dependency Updates
+
+* `mui-feedback` updated to version `0.2.35`
 ## [0.2.35](https://github.com/Availity/element/compare/@availity/mui-page-header@0.2.34...@availity/mui-page-header@0.2.35) (2024-06-14)
 
 ### Dependency Updates